Log warnings and errors to Windows event log
