+++ clcache/clcache_lib/setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="clcache-lib",
-    version="4.4.4aa",
+    version="4.4.4",
     author="Various",
     author_email="Various",
     packages=find_packages(exclude=('clcache_lib',)),
